+++ reqwest-middleware/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added `ClientBuilder::with_final_init` and `ClientBuilder::with_arc_final_init`
+
 ## [0.4.2] - 2025-04-08
 
 ### Added